+++ src/ConnectKafkaFake.cpp
@@ -29,10 +29,10 @@ int ConnectKafkaFake::getNumberOfTopicPartitions(std::string TopicName) {
   return 1;
 }
 
-void ConnectKafkaFake::subscribeAtOffset(int64_t Offset,
+void ConnectKafkaFake::subscribeAtOffset(int32_t Offset,
                                          std::string TopicName) {}
 
-void ConnectKafkaFake::subscribeToLastNMessages(int64_t NMessages,
+void ConnectKafkaFake::subscribeToLastNMessages(int32_t NMessages,
                                                 const std::string &TopicName,
                                                 int Partition) {}
