+++ src/UserArgumentsStruct.h
@@ -8,7 +8,7 @@ struct UserArgumentStruct {
   // value -2 signalises that the variable hasn't been set by a user and won't
   // be used
   std::int16_t GoBack = -2;
-  std::int64_t OffsetToStart = -2;
+  std::int32_t OffsetToStart = -2;
 
   int PartitionToConsume = -1;
